Split home.utils into separate modules
